+++ src/irmin-pack/pack.ml
@@ -42,8 +42,7 @@ module Cache (K : Irmin.Type.S) = Lru.Make (struct
   let equal (x : t) (y : t) = Irmin.Type.equal K.t x y
 end)
 
-let with_cache = IO.with_cache
-
+module IO_cache = IO.Cache
 module IO = IO.Unix
 
 module File (Index : Pack_index.S) (K : Irmin.Hash.S with type t = Index.key) =
@@ -76,11 +75,13 @@ struct
     let root = Filename.dirname file in
     let lock = Lwt_mutex.create () in
     let dict = Dict.v ~fresh ~readonly root in
-    let block = IO.v ~version:current_version ~fresh ~readonly file in
+    let block = IO.v ~version:(Some current_version) ~fresh ~readonly file in
     { block; index; lock; dict; open_instances = 1 }
 
-  let (`Staged v) =
-    with_cache ~clear ~valid ~v:(fun index -> unsafe_v ~index) "store.pack"
+  let IO_cache.{ v } =
+    IO_cache.memoize ~clear ~valid
+      ~v:(fun index -> unsafe_v ~index)
+      "store.pack"
 
   type key = K.t
 
@@ -308,7 +309,7 @@ struct
         on_generation_change ();
         IO.close t.pack.block;
         let block =
-          IO.v ~fresh:false ~version:current_version ~readonly:true
+          IO.v ~fresh:false ~version:(Some current_version) ~readonly:true
             (IO.name t.pack.block)
         in
         t.pack.block <- block;