+++ test/irmin-pack/common.ml
@@ -12,6 +12,12 @@ let rm_dir root =
     let _ = Sys.command cmd in
     ())
 
+module Conf = struct
+  let entries = 32
+
+  let stable_hash = 256
+end
+
 module S = struct
   include Irmin.Contents.String
 
@@ -90,4 +96,128 @@ struct
   let close index pack =
     Index.close index;
     Pack.close pack
+end
+
+module Alcotest = struct
+  include Alcotest
+
+  (** TODO: upstream this to Alcotest *)
+  let check_raises_lwt msg exn (type a) (f : unit -> a Lwt.t) =
+    Lwt.catch
+      (fun x ->
+        f x >>= fun (_ : a) ->
+        Alcotest.failf
+          "Fail %s: expected function to raise %s, but it returned instead." msg
+          (Printexc.to_string exn))
+      (function
+        | e when e = exn -> Lwt.return_unit
+        | e ->
+            Alcotest.failf
+              "Fail %s: expected function to raise %s, but it raised %s \
+               instead."
+              msg (Printexc.to_string exn) (Printexc.to_string e))
+end
+
+module Filename = struct
+  include Filename
+
+  (** Extraction from OCaml for pre-4.10 compatibility *)
+
+  let generic_quote quotequote s =
+    let l = String.length s in
+    let b = Buffer.create (l + 20) in
+    Buffer.add_char b '\'';
+    for i = 0 to l - 1 do
+      if s.[i] = '\'' then Buffer.add_string b quotequote
+      else Buffer.add_char b s.[i]
+    done;
+    Buffer.add_char b '\'';
+    Buffer.contents b
+
+  let quote_command =
+    match Sys.os_type with
+    | "Win32" ->
+        let quote s =
+          let l = String.length s in
+          let b = Buffer.create (l + 20) in
+          Buffer.add_char b '\"';
+          let rec loop i =
+            if i = l then Buffer.add_char b '\"'
+            else
+              match s.[i] with
+              | '\"' -> loop_bs 0 i
+              | '\\' -> loop_bs 0 i
+              | c ->
+                  Buffer.add_char b c;
+                  loop (i + 1)
+          and loop_bs n i =
+            if i = l then (
+              Buffer.add_char b '\"';
+              add_bs n)
+            else
+              match s.[i] with
+              | '\"' ->
+                  add_bs ((2 * n) + 1);
+                  Buffer.add_char b '\"';
+                  loop (i + 1)
+              | '\\' -> loop_bs (n + 1) (i + 1)
+              | _ ->
+                  add_bs n;
+                  loop i
+          and add_bs n =
+            for _j = 1 to n do
+              Buffer.add_char b '\\'
+            done
+          in
+          loop 0;
+          Buffer.contents b
+        in
+        let quote_cmd s =
+          let b = Buffer.create (String.length s + 20) in
+          String.iter
+            (fun c ->
+              match c with
+              | '(' | ')' | '!' | '^' | '%' | '\"' | '<' | '>' | '&' | '|' ->
+                  Buffer.add_char b '^';
+                  Buffer.add_char b c
+              | _ -> Buffer.add_char b c)
+            s;
+          Buffer.contents b
+        in
+        let quote_cmd_filename f =
+          if String.contains f '\"' || String.contains f '%' then
+            failwith ("Filename.quote_command: bad file name " ^ f)
+          else if String.contains f ' ' then "\"" ^ f ^ "\""
+          else f
+        in
+        fun cmd ?stdin ?stdout ?stderr args ->
+          String.concat ""
+            [
+              "\"";
+              quote_cmd_filename cmd;
+              " ";
+              quote_cmd (String.concat " " (List.map quote args));
+              (match stdin with
+              | None -> ""
+              | Some f -> " <" ^ quote_cmd_filename f);
+              (match stdout with
+              | None -> ""
+              | Some f -> " >" ^ quote_cmd_filename f);
+              (match stderr with
+              | None -> ""
+              | Some f ->
+                  if stderr = stdout then " 2>&1"
+                  else " 2>" ^ quote_cmd_filename f);
+              "\"";
+            ]
+    | _ -> (
+        let quote = generic_quote "'\\''" in
+        fun cmd ?stdin ?stdout ?stderr args ->
+          String.concat " " (List.map quote (cmd :: args))
+          ^ (match stdin with None -> "" | Some f -> " <" ^ quote f)
+          ^ (match stdout with None -> "" | Some f -> " >" ^ quote f)
+          ^
+          match stderr with
+          | None -> ""
+          | Some f -> if stderr = stdout then " 2>&1" else " 2>" ^ quote f)
 end