+++ test/irmin-pack/test.ml
@@ -15,5 +15,6 @@
  *)
 
 let () =
-  Irmin_test.Store.run "irmin-pack" ~misc:Test_pack.misc
+  Irmin_test.Store.run "irmin-pack"
+    ~misc:(Test_pack.misc @ [ ("utils", Test_utils.tests) ])
     [ (`Quick, Test_pack.suite) ]