+++ test/irmin-pack/test_pack.ml
@@ -593,4 +593,5 @@ let misc =
     ("pack-files", Pack.tests);
     ("branch-files", Branch.tests);
     ("instances", Multiple_instances.tests);
+    ("migration", Migration.tests);
   ]