+++ CHANGES.md
@@ -31,7 +31,11 @@
 - **irmin-pack**:
   - Added `index_throttle` option to `Irmin_pack.config`, which exposes the
     memory throttle feature of `Index` in `Irmin-Pack`. (#1049, @icristescu)
+
   - Added `Pack.clear` and `Dict.clear` (#1047, @icristescu, @CraigFe, @samoht)
+
+  - Added a `migrate` function for upgrading stores with old formats (#TODO,
+    @icristescu, @CraigFe)
 
 #### Changed
 
@@ -44,8 +48,10 @@
 - **irmin-pack**:
   - `sync` has to be called by the read-only instance to synchronise with the
     files on disk. (#1008, @icristescu)
+
   - Renamed `sync` to `flush` for the operation that flushes to disk all buffers
     of a read-write instance. (#1008, @icristescu)
+
   - Changed the format of headers for the files on disk to include a generation
     number. Version 1 of irmin-pack was used for the previous format, version 2
     is used with the new format. (#1047, @icristescu, @CraigFe, @samoht)